+++ types/security.d.ts
@@ -2,3 +2,6 @@ export function generateUID(length?: number): string;
 export function hashString(str: string): string;
 export function base64Encode(str: string): string;
 export function base64Decode(str: string): string;
+export function generateOTP(length?: number): string;
+export function encryptString(str: string, key: string): string;
+export function decryptString(str: string, key: string): string;